requirements.txt
flask
gunicorn
numpy
markdown2
python-frontmatter